+++ main-server/src/tendermint/ndid.js
@@ -953,7 +953,7 @@ export async function getIdpNodes({
   agent,
   min_ial,
   min_aal,
-  on_the_fly_support,
+  supported_feature_list,
   node_id_list,
   supported_request_message_type_list,
   mode_list,
@@ -978,7 +978,7 @@ export async function getIdpNodes({
       agent,
       min_ial,
       min_aal,
-      on_the_fly_support,
+      supported_feature_list,
       node_id_list,
       supported_request_message_type_list,
       mode_list,
@@ -1844,6 +1844,69 @@ export async function isSuppressedIdentityModificationNotificationNode({
     throw new CustomError({
       message:
         'Cannot check is suppressed identity modification notification node from blockchain',
+      cause: error,
+    });
+  }
+}
+
+export async function addAllowedNodeSupportedFeature(
+  { name },
+  nodeId,
+  callbackFnName,
+  callbackAdditionalArgs,
+  saveForRetryOnChainDisabled
+) {
+  try {
+    await tendermint.transact({
+      nodeId,
+      fnName: 'AddAllowedNodeSupportedFeature',
+      params: { name },
+      callbackFnName,
+      callbackAdditionalArgs,
+      saveForRetryOnChainDisabled,
+    });
+  } catch (error) {
+    throw new CustomError({
+      message: 'Cannot add allowed node supported feature',
+      cause: error,
+    });
+  }
+}
+
+export async function removeAllowedNodeSupportedFeature(
+  { name },
+  nodeId,
+  callbackFnName,
+  callbackAdditionalArgs,
+  saveForRetryOnChainDisabled
+) {
+  try {
+    await tendermint.transact({
+      nodeId,
+      fnName: 'RemoveAllowedNodeSupportedFeature',
+      params: { name },
+      callbackFnName,
+      callbackAdditionalArgs,
+      saveForRetryOnChainDisabled,
+    });
+  } catch (error) {
+    throw new CustomError({
+      message: 'Cannot remove allowed node supported feature',
+      cause: error,
+    });
+  }
+}
+
+export async function getAllowedNodeSupportedFeatureList({ prefix } = {}) {
+  try {
+    const result = await tendermint.query(
+      'GetAllowedNodeSupportedFeatureList',
+      { prefix }
+    );
+    return result;
+  } catch (error) {
+    throw new CustomError({
+      message: 'Cannot get allowed node supported feature list',
       cause: error,
     });
   }