+++ CHANGELOG.md
@@ -65,6 +65,22 @@ BREAKING CHANGES:
       - Add `encryption_public_key`
       - Add `encryption_key_algorithm`
       - Add `encryption_algorithm`
+  - Remove query string parameter `on_the_fly_support` from GET `/utility/idp`
+  - Add query string parameter `supported_feature_list` (list of strings separated by `,`) to GET `/utility/idp`
+  - Change response body JSON schema of GET `/utility/idp`.
+    - Remove `on_the_fly_support` (boolean) property.
+    - Add `supported_feature_list` (array of strings) property.
+  - Remove query string parameter `on_the_fly_support` from GET `/utility/idp/:namespace/:identifier`
+  - Add query string parameter `supported_feature_list` (list of strings separated by `,`) to GET `/utility/idp/:namespace/:identifier`
+  - Change response body JSON schema of GET `/utility/idp/:namespace/:identifier`.
+    - Remove `on_the_fly_support` (boolean) property.
+    - Add `supported_feature_list` (array of strings) property.
+  - Change response body JSON schema of GET `/utility/nodes/:node_id`.
+    - Remove `on_the_fly_support` (boolean) property.
+    - Add `supported_feature_list` (array of strings) property.
+  - NDID only APIs
+    - Remove `on_the_fly_support` property for registering node (POST `/ndid/register_node`) and updating node (POST `/ndid/update_node`).
+    - Add `supported_feature_list` property for registering node (POST `/ndid/register_node`) and updating node (POST `/ndid/update_node`).
 
 FEATURES:
 
@@ -89,6 +105,10 @@ FEATURES:
   - `RSAES_OAEP_SHA_256`
 - API version 6.0
   - New API: GET `/utility/nodes/:node_id/public_keys`: Get node public key list (including previous key versions).
+  - New API: GET `/utility/node_supported_features`: Get valid node supported feature list.
+  - NDID only APIs
+    - New API: POST `/ndid/add_allowed_node_supported_feature`: Add allowed node supported feature
+    - New API: POST `/ndid/remove_allowed_node_supported_feature`: Remove allowed node supported feature
 - Add environment variables.
   - `SIGNING_PRIVATE_KEY_PATH`: Path to node's private key for signing [Default: use pre-generated development key in development mode]
   - `SIGNING_PRIVATE_KEY_PASSPHRASE`: Passphrase for node's private key for signing